backend/requirements.txt
# ==========================================
# Yetria Career Guidance Platform
# Backend Dependencies
# ==========================================

# ===== WEB FRAMEWORK =====
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ===== DATABASE =====
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL adapter (EKLENDI)

# ===== AUTHENTICATION & SECURITY =====
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1

# ===== DATA VALIDATION =====
pydantic[email]==2.5.0
email-validator==2.3.0

# ===== UTILITIES =====
python-dotenv==1.0.0

# ===== DATA PROCESSING =====
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0

# ===== MACHINE LEARNING MODELS =====
lightgbm==4.0.0
catboost==1.2.2
# xgboost==1.7.6  # REMOVED: Not used in codebase

# ===== MODEL PERSISTENCE =====
joblib==1.3.2

# ===== VISUALIZATION (for ML training) =====
matplotlib==3.7.2
seaborn==0.12.2

# ===== DEVELOPMENT & TESTING (optional) =====
# Uncomment these if you need testing tools
# pytest==7.4.3
# pytest-asyncio==0.21.1
# httpx==0.25.2